+++ packages/pointer_interceptor/pointer_interceptor_ios/example/integration_test/pointer_interceptor_test.dart
@@ -8,5 +8,8 @@ import 'package:integration_test/integration_test.dart';
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
+  // Integration tests for iOS requires a native platform view and thus
+  // can not be tested in Dart. These tests can instead
+  // be found in the XCUITests
   testWidgets('placeholder test', (WidgetTester tester) async {});
 }