+++ packages/pointer_interceptor/pointer_interceptor_web/CHANGELOG.md
@@ -1,4 +1,4 @@
 ## 0.10.0
 
-* Initial release.
+* Moved web implementation to own package.